Add support to multiline arrays on dockerfiles
