+++ frontend/dockerfile/parser/parser.go
@@ -79,10 +79,11 @@ func (node *Node) AddChild(child *Node, startLine, endLine int) {
 }
 
 var (
-	dispatch           map[string]func(string, *Directive) (*Node, map[string]bool, error)
-	tokenWhitespace    = regexp.MustCompile(`[\t\v\f\r ]+`)
-	tokenEscapeCommand = regexp.MustCompile(`^#[ \t]*escape[ \t]*=[ \t]*(?P<escapechar>.).*$`)
-	tokenComment       = regexp.MustCompile(`^#.*$`)
+	dispatch                 map[string]func(string, *Directive) (*Node, map[string]bool, error)
+	tokenWhitespace          = regexp.MustCompile(`[\t\v\f\r ]+`)
+	tokenEscapeCommand       = regexp.MustCompile(`^#[ \t]*escape[ \t]*=[ \t]*(?P<escapechar>.).*$`)
+	tokenComment             = regexp.MustCompile(`^#.*$`)
+	lineJSONArrayContinuator = regexp.MustCompile(`[^"]*\[[^\]]*$`)
 )
 
 // DefaultEscapeToken is the default escape token
@@ -91,10 +92,10 @@ const DefaultEscapeToken = '\\'
 // Directive is the structure used during a build run to hold the state of
 // parsing directives.
 type Directive struct {
-	escapeToken           rune           // Current escape token
-	lineContinuationRegex *regexp.Regexp // Current line continuation regex
-	processingComplete    bool           // Whether we are done looking for directives
-	escapeSeen            bool           // Whether the escape directive has been seen
+	escapeToken        rune           // Current escape token
+	lineEscapeRegex    *regexp.Regexp // Current line escape regex
+	processingComplete bool           // Whether we are done looking for directives
+	escapeSeen         bool           // Whether the escape directive has been seen
 }
 
 // setEscapeToken sets the default token for escaping characters in a Dockerfile.
@@ -103,7 +104,7 @@ func (d *Directive) setEscapeToken(s string) error {
 		return fmt.Errorf("invalid ESCAPE '%s'. Must be ` or \\", s)
 	}
 	d.escapeToken = rune(s[0])
-	d.lineContinuationRegex = regexp.MustCompile(`\` + s + `[ \t]*$`)
+	d.lineEscapeRegex = regexp.MustCompile(`\` + s + `[ \t]*$`)
 	return nil
 }
 
@@ -234,7 +235,7 @@ func Parse(rwc io.Reader) (*Result, error) {
 		currentLine++
 
 		startLine := currentLine
-		line, isEndOfLine := trimContinuationCharacter(string(bytesRead), d)
+		line, isEndOfLine := continuateLine(string(bytesRead), d)
 		if isEndOfLine && line == "" {
 			continue
 		}
@@ -257,8 +258,7 @@ func Parse(rwc io.Reader) (*Result, error) {
 			}
 
 			continuationLine := string(bytesRead)
-			continuationLine, isEndOfLine = trimContinuationCharacter(continuationLine, d)
-			line += continuationLine
+			line, isEndOfLine = continuateLine(line+continuationLine, d)
 		}
 
 		if hasEmptyContinuationLine {
@@ -305,11 +305,15 @@ func isEmptyContinuationLine(line []byte) bool {
 
 var utf8bom = []byte{0xEF, 0xBB, 0xBF}
 
-func trimContinuationCharacter(line string, d *Directive) (string, bool) {
-	if d.lineContinuationRegex.MatchString(line) {
-		line = d.lineContinuationRegex.ReplaceAllString(line, "")
+func continuateLine(line string, d *Directive) (string, bool) {
+	if d.lineEscapeRegex.MatchString(line) {
+		line = d.lineEscapeRegex.ReplaceAllString(line, "")
 		return line, false
 	}
+	if lineJSONArrayContinuator.MatchString(line) {
+		return line, false
+	}
+
 	return line, true
 }
